+++ src/ert/parsing/__init__.py
@@ -1,6 +1,5 @@
 from .config_errors import ConfigValidationError, ConfigWarning
 from .config_keywords import ConfigKeys
-from .config_parsing_helpers import parse_arg_types_list
 from .config_schema import (
     SchemaItemType,
     init_site_config_schema,
@@ -24,7 +23,6 @@ __all__ = [
     "SchemaItemType",
     "init_site_config_schema",
     "init_user_config_schema",
-    "parse_arg_types_list",
     "ConfigDict",
     "ExtJobKeys",
     "init_ext_job_schema",
